For loops and arrays on the stack Added: - For loops that work with slices.   Use @reverse annotation to iterate in reverse. - Arrays can be made on the stack.   A slice is used to refer to the array. Example: arr: i32[32] - char[] is an alternative for Slice<char>.   This (should) works with all types. Fixes: - Index operator did not work for assignments before but now they do.   Example: ptr[3] = 23

+++ include/BetBat/Generator.h
@@ -22,7 +22,7 @@ struct GenInfo {
     // Negative value to make some space for values
     // Positive to remove values
     // like push and pop but with a size
-    void addStackSpace(i16 size);
+    void addStackSpace(i32 size);
     int saveStackMoment();
     void restoreStackMoment(int moment);
